Do not return error from xcgroup_delete if cgroup already gone
Some cgroup code would retry continuously if somehow a cgroup
didn't exist when xcgroup_delete() was called (See for instance
proctrack/cgroup's slurm_container_plugin_wait()). To fix this
for all callers, return SUCCESS from xcgroup_delete() if rmdir(2)
return ENOENT, since the cgroup is already deleted.

+++ src/common/xcgroup.c
@@ -504,7 +504,11 @@ int xcgroup_load(xcgroup_ns_t* cgns, xcgroup_t* cg, char* uri)
 
 int xcgroup_delete(xcgroup_t* cg)
 {
-	if (rmdir(cg->path) < 0) {
+	/*
+	 *  Simply delete cgroup with rmdir(2). If cgroup doesn't
+	 *   exist, do not propagate error back to caller.
+	 */
+	if ((rmdir(cg->path) < 0) && (errno != ENOENT)) {
 		verbose ("xcgroup: rmdir(%s): %m", cg->path);
 		return XCGROUP_ERROR;
 	}
